Added various uncommitted stuff of the last weeks :-(

+++ includes/modules/xsell_products.php
@@ -14,7 +14,7 @@ Released under the GNU General Public License
 if ($HTTP_GET_VARS['products_id']) {
 $xsell_query = tep_db_query("select distinct p.products_id, p.products_image, pd.products_name from " . TABLE_PRODUCTS_XSELL . " xp, " . TABLE_PRODUCTS . " p left join ". TABLE_PRODUCTS_TO_CATEGORIES. " p2c using (products_id) left join ". TABLE_CATEGORIES ." c using (categories_id), " . TABLE_PRODUCTS_DESCRIPTION . " pd where xp.products_id = '" . $HTTP_GET_VARS['products_id'] . "' and xp.xsell_id = p.products_id and p.products_id = pd.products_id and pd.language_id = '" . $languages_id . "' and c.categories_status=1 and p.products_status = '1' order by xp.products_id asc limit " . MAX_DISPLAY_ALSO_PURCHASED);
 $num_products_xsell = tep_db_num_rows($xsell_query);
-if ($num_products_xsell >= MIN_DISPLAY_ALSO_PURCHASED) {
+if ($num_products_xsell >= (int)MIN_DISPLAY_ALSO_PURCHASED) {
 ?>
 <!-- xsell_products //-->
 <?php
@@ -40,9 +40,6 @@ if ($num_products_xsell >= MIN_DISPLAY_ALSO_PURCHASED) {
 ?>
 <!-- xsell_products_eof //-->
 <br />
-
-
-
 <?php
 		}
 	}
